QS: added Memory usage bar to the QS Panel
Supports top/bottom position.

This was moved from the Recents panel tweaks as those are already obsolete
(and thus removed within this commit as well).
The recents implementation got complicated since it does not have to be a system component anymore.
Launchers can implement a recents component themselves which
system will use instead of one in SystemUI. This makes it hard to provide
any recents panel related tweaks.