Add lockscreen carrier text to quicksettings
I'm not 100% sure if this is where this commit belongs, but
I didn't want to refactor the carrier text out of the ModLockscreen file
without first checking with someone.

This commit will add the custom carrier text a user specifies in the
lockscreen preference, to the quick settings carrier text view. It does
this by hooking a method called `setCarrierText()` in systemui package.
From what I can see this seems to be a pretty good place to make the
hook because not that many calls are made to it.

See here for the hooked method in the android source:
https://github.com/aosp-mirror/platform_frameworks_base/blob/1e4b31dd5dbbc3ef24155e8076edbfd80156f820/packages/SystemUI/src/com/android/systemui/qs/QSCarrier.java#L111:17